refactor: consolidate CLI utilities from @inquirerer/utils
- Update package.json dependencies to use inquirerer@4.0.0, @inquirerer/utils@3.0.0, genomic@5.0.0
- Replace local extractFirst() with @inquirerer/utils export
- Update create-gen-app imports to use genomic for scaffolding
- Delete duplicated argv.ts files from packages/cli and pgpm/cli
- Update direct imports to use utils index re-exports
