+++ packages/cli/package.json
@@ -55,7 +55,8 @@
     "@pgpmjs/server-utils": "workspace:^",
     "@pgpmjs/types": "workspace:^",
     "find-and-require-package-json": "^0.8.2",
-    "inquirerer": "^2.4.0",
+    "@inquirerer/utils": "^3.0.0",
+    "inquirerer": "^4.0.0",
     "js-yaml": "^4.1.0",
     "minimist": "^1.2.8",
     "pg-cache": "workspace:^",
