+++ pgpm/cli/src/commands/add.ts
@@ -3,7 +3,7 @@ import { CLIOptions, Inquirerer } from 'inquirerer';
 import { ParsedArgs } from 'minimist';
 import * as path from 'path';
 
-import { extractFirst } from '../utils/argv';
+import { extractFirst } from '../utils';
 
 const addUsageText = `
 Add Command:
